move transaction related functionality to data_model/transaction module
