+++ data_model/src/events.rs
@@ -179,19 +179,12 @@ pub mod data {
 
 /// Pipeline events.
 pub mod pipeline {
-    use std::{
-        error::Error as StdError,
-        fmt::{Display, Formatter, Result as FmtResult},
-    };
-
-    use iroha_crypto::{Hash, SignatureVerificationFail};
+    pub use crate::transaction::RejectionReason as PipelineRejectionReason;
+    use iroha_crypto::Hash;
     use iroha_macro::FromVariant;
     use iroha_schema::prelude::*;
     use parity_scale_codec::{Decode, Encode};
     use serde::{Deserialize, Serialize};
-    use thiserror::Error;
-
-    use crate::{isi::Instruction, transaction::Payload};
 
     /// Event filter.
     #[derive(Debug, Decode, Encode, Deserialize, Serialize, Copy, Clone, IntoSchema)]
@@ -256,152 +249,6 @@ pub mod pipeline {
         Transaction,
     }
 
-    /// Transaction was reject because it doesn't satisfy signature condition
-    #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Decode, Encode, IntoSchema)]
-    pub struct UnsatisfiedSignatureConditionFail {
-        /// Reason why signature condition failed
-        pub reason: String,
-    }
-
-    impl Display for UnsatisfiedSignatureConditionFail {
-        fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-            write!(
-                f,
-                "Failed to verify signature condition specified in the account: {}",
-                self.reason,
-            )
-        }
-    }
-
-    impl StdError for UnsatisfiedSignatureConditionFail {}
-
-    /// Transaction was rejected because of one of its instructions failing.
-    #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Decode, Encode, IntoSchema)]
-    pub struct InstructionExecutionFail {
-        /// Instruction which execution failed
-        pub instruction: Instruction,
-        /// Error which happened during execution
-        pub reason: String,
-    }
-
-    impl Display for InstructionExecutionFail {
-        fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-            use Instruction::*;
-            let type_ = match self.instruction {
-                Burn(_) => "burn",
-                Fail(_) => "fail",
-                If(_) => "if",
-                Mint(_) => "mint",
-                Pair(_) => "pair",
-                Register(_) => "register",
-                Sequence(_) => "sequence",
-                Transfer(_) => "transfer",
-                Unregister(_) => "unregister",
-                SetKeyValue(_) => "set key-value pair",
-                RemoveKeyValue(_) => "remove key-value pair",
-                Grant(_) => "grant",
-            };
-            write!(
-                f,
-                "Failed to execute instruction of type {}: {}",
-                type_, self.reason
-            )
-        }
-    }
-    impl StdError for InstructionExecutionFail {}
-
-    /// Transaction was reject because of low authority
-    #[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Decode, Encode, IntoSchema)]
-    pub struct NotPermittedFail {
-        /// Reason of failure
-        pub reason: String,
-    }
-
-    impl Display for NotPermittedFail {
-        fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
-            write!(f, "Action not permitted: {}", self.reason)
-        }
-    }
-
-    impl StdError for NotPermittedFail {}
-
-    /// The reason for rejecting transaction which happened because of new blocks.
-    #[derive(
-        Debug,
-        Clone,
-        Copy,
-        Eq,
-        PartialEq,
-        Serialize,
-        Deserialize,
-        Decode,
-        Encode,
-        FromVariant,
-        Error,
-        IntoSchema,
-    )]
-    pub enum BlockRejectionReason {
-        /// Block was rejected during consensus.
-        //TODO: store rejection reasons for blocks?
-        #[error("Block was rejected during consensus.")]
-        ConsensusBlockRejection,
-    }
-
-    /// The reason for rejecting transaction which happened because of transaction.
-    #[derive(
-        Debug,
-        Clone,
-        Eq,
-        PartialEq,
-        Serialize,
-        Deserialize,
-        Decode,
-        Encode,
-        FromVariant,
-        Error,
-        IntoSchema,
-    )]
-    pub enum TransactionRejectionReason {
-        /// Insufficient authorisation.
-        #[error("Transaction rejected due to insufficient authorisation")]
-        NotPermitted(#[source] NotPermittedFail),
-        /// Failed to verify signature condition specified in the account.
-        #[error("Transaction rejected due to an unsatisfied signature condition")]
-        UnsatisfiedSignatureCondition(#[source] UnsatisfiedSignatureConditionFail),
-        /// Failed to execute instruction.
-        #[error("Transaction rejected due to failure in instruction execution")]
-        InstructionExecution(#[source] InstructionExecutionFail),
-        /// Failed to verify signatures.
-        #[error("Transaction rejected due to failed signature verification")]
-        SignatureVerification(#[source] SignatureVerificationFail<Payload>),
-        /// Genesis account can sign only transactions in the genesis block.
-        #[error("The genesis account can only sign transactions in the genesis block.")]
-        UnexpectedGenesisAccountSignature,
-    }
-
-    /// The reason for rejecting pipeline entity such as transaction or block.
-    #[derive(
-        Debug,
-        Clone,
-        Eq,
-        PartialEq,
-        Serialize,
-        Deserialize,
-        Decode,
-        Encode,
-        FromVariant,
-        Error,
-        IntoSchema,
-    )]
-    pub enum RejectionReason {
-        /// The reason for rejecting the block.
-        #[error("Block was rejected")]
-        Block(#[source] BlockRejectionReason),
-        /// The reason for rejecting transaction.
-        #[error("Transaction was rejected")]
-        Transaction(#[source] TransactionRejectionReason),
-    }
-
     /// Entity type to filter events.
     #[derive(Debug, Decode, Encode, Deserialize, Serialize, Eq, PartialEq, Clone, IntoSchema)]
     pub struct Event {
@@ -432,7 +279,7 @@ pub mod pipeline {
         /// Entity has been seen in blockchain, but has not passed validation.
         Validating,
         /// Entity was rejected in one of the validation stages.
-        Rejected(RejectionReason),
+        Rejected(PipelineRejectionReason),
         /// Entity has passed validation.
         Committed,
     }
@@ -440,10 +287,8 @@ pub mod pipeline {
     /// Exports common structs and enums from this module.
     pub mod prelude {
         pub use super::{
-            BlockRejectionReason, EntityType as PipelineEntityType, Event as PipelineEvent,
-            EventFilter as PipelineEventFilter, InstructionExecutionFail, NotPermittedFail,
-            RejectionReason as PipelineRejectionReason, Status as PipelineStatus,
-            TransactionRejectionReason, UnsatisfiedSignatureConditionFail,
+            EntityType as PipelineEntityType, Event as PipelineEvent,
+            EventFilter as PipelineEventFilter, PipelineRejectionReason, Status as PipelineStatus,
         };
     }
 
@@ -451,8 +296,7 @@ pub mod pipeline {
     mod tests {
         #![allow(clippy::restriction)]
 
-        use RejectionReason::*;
-        use TransactionRejectionReason::*;
+        use crate::transaction::{RejectionReason::*, TransactionRejectionReason::*};
 
         use super::*;
