+++ data_model/src/transaction.rs
@@ -3,17 +3,22 @@
 use std::{cmp::Ordering, collections::BTreeSet, vec::IntoIter as VecIter};
 
 use eyre::{eyre, Result};
-use iroha_crypto::{HashOf, KeyPair, SignatureOf, SignaturesOf};
+use iroha_crypto::{HashOf, KeyPair, SignatureOf, SignatureVerificationFail, SignaturesOf};
+use iroha_macro::FromVariant;
 use iroha_schema::IntoSchema;
 use iroha_version::{declare_versioned, declare_versioned_with_scale, version, version_with_scale};
 use parity_scale_codec::{Decode, Encode};
 use serde::{Deserialize, Serialize};
+use std::{
+    error::Error as StdError,
+    fmt::{Display, Formatter, Result as FmtResult},
+};
+use thiserror::Error;
 #[cfg(feature = "warp")]
 use warp::{reply::Response, Reply};
 
 use crate::{
-    account::Account, current_time, isi::Instruction, metadata::UnlimitedMetadata,
-    prelude::TransactionRejectionReason, Identifiable,
+    account::Account, current_time, isi::Instruction, metadata::UnlimitedMetadata, Identifiable,
 };
 
 /// Default maximum number of instructions and expressions per transaction
@@ -96,7 +101,7 @@ declare_versioned!(
     Clone,
     PartialEq,
     Eq,
-    iroha_macro::FromVariant,
+    FromVariant,
     IntoSchema,
 );
 
@@ -268,7 +273,7 @@ impl Txn for Transaction {
     }
 }
 
-declare_versioned_with_scale!(VersionedPendingTransactions 1..2, iroha_macro::FromVariant, Clone, Debug);
+declare_versioned_with_scale!(VersionedPendingTransactions 1..2, FromVariant, Clone, Debug);
 
 impl VersionedPendingTransactions {
     /// Converts from `&VersionedPendingTransactions` to V1 reference
@@ -372,7 +377,7 @@ impl PartialOrd for TransactionValue {
     }
 }
 
-declare_versioned_with_scale!(VersionedValidTransaction 1..2, Debug, Clone, iroha_macro::FromVariant, IntoSchema);
+declare_versioned_with_scale!(VersionedValidTransaction 1..2, Debug, Clone, FromVariant, IntoSchema);
 
 impl VersionedValidTransaction {
     /// Converts from `&VersionedValidTransaction` to V1 reference
@@ -426,7 +431,7 @@ impl Txn for ValidTransaction {
     }
 }
 
-declare_versioned!(VersionedRejectedTransaction 1..2, Clone, Debug, PartialEq, Eq, iroha_macro::FromVariant, IntoSchema);
+declare_versioned!(VersionedRejectedTransaction 1..2, Clone, Debug, PartialEq, Eq, FromVariant, IntoSchema);
 
 impl VersionedRejectedTransaction {
     /// Converts from `&VersionedRejectedTransaction` to V1 reference
@@ -483,10 +488,158 @@ impl Txn for RejectedTransaction {
     }
 }
 
+/// Transaction was reject because it doesn't satisfy signature condition
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Decode, Encode, IntoSchema)]
+pub struct UnsatisfiedSignatureConditionFail {
+    /// Reason why signature condition failed
+    pub reason: String,
+}
+
+impl Display for UnsatisfiedSignatureConditionFail {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        write!(
+            f,
+            "Failed to verify signature condition specified in the account: {}",
+            self.reason,
+        )
+    }
+}
+
+impl StdError for UnsatisfiedSignatureConditionFail {}
+
+/// Transaction was rejected because of one of its instructions failing.
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Decode, Encode, IntoSchema)]
+pub struct InstructionExecutionFail {
+    /// Instruction which execution failed
+    pub instruction: Instruction,
+    /// Error which happened during execution
+    pub reason: String,
+}
+
+impl Display for InstructionExecutionFail {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        use Instruction::*;
+        let type_ = match self.instruction {
+            Burn(_) => "burn",
+            Fail(_) => "fail",
+            If(_) => "if",
+            Mint(_) => "mint",
+            Pair(_) => "pair",
+            Register(_) => "register",
+            Sequence(_) => "sequence",
+            Transfer(_) => "transfer",
+            Unregister(_) => "unregister",
+            SetKeyValue(_) => "set key-value pair",
+            RemoveKeyValue(_) => "remove key-value pair",
+            Grant(_) => "grant",
+        };
+        write!(
+            f,
+            "Failed to execute instruction of type {}: {}",
+            type_, self.reason
+        )
+    }
+}
+impl StdError for InstructionExecutionFail {}
+
+/// Transaction was reject because of low authority
+#[derive(Debug, Clone, Eq, PartialEq, Serialize, Deserialize, Decode, Encode, IntoSchema)]
+pub struct NotPermittedFail {
+    /// Reason of failure
+    pub reason: String,
+}
+
+impl Display for NotPermittedFail {
+    fn fmt(&self, f: &mut Formatter<'_>) -> FmtResult {
+        write!(f, "Action not permitted: {}", self.reason)
+    }
+}
+
+impl StdError for NotPermittedFail {}
+
+/// The reason for rejecting transaction which happened because of new blocks.
+#[derive(
+    Debug,
+    Clone,
+    Copy,
+    Eq,
+    PartialEq,
+    Serialize,
+    Deserialize,
+    Decode,
+    Encode,
+    FromVariant,
+    Error,
+    IntoSchema,
+)]
+pub enum BlockRejectionReason {
+    /// Block was rejected during consensus.
+    //TODO: store rejection reasons for blocks?
+    #[error("Block was rejected during consensus.")]
+    ConsensusBlockRejection,
+}
+
+/// The reason for rejecting transaction which happened because of transaction.
+#[derive(
+    Debug,
+    Clone,
+    Eq,
+    PartialEq,
+    Serialize,
+    Deserialize,
+    Decode,
+    Encode,
+    FromVariant,
+    Error,
+    IntoSchema,
+)]
+pub enum TransactionRejectionReason {
+    /// Insufficient authorisation.
+    #[error("Transaction rejected due to insufficient authorisation")]
+    NotPermitted(#[source] NotPermittedFail),
+    /// Failed to verify signature condition specified in the account.
+    #[error("Transaction rejected due to an unsatisfied signature condition")]
+    UnsatisfiedSignatureCondition(#[source] UnsatisfiedSignatureConditionFail),
+    /// Failed to execute instruction.
+    #[error("Transaction rejected due to failure in instruction execution")]
+    InstructionExecution(#[source] InstructionExecutionFail),
+    /// Failed to verify signatures.
+    #[error("Transaction rejected due to failed signature verification")]
+    SignatureVerification(#[source] SignatureVerificationFail<Payload>),
+    /// Genesis account can sign only transactions in the genesis block.
+    #[error("The genesis account can only sign transactions in the genesis block.")]
+    UnexpectedGenesisAccountSignature,
+}
+
+/// The reason for rejecting pipeline entity such as transaction or block.
+#[derive(
+    Debug,
+    Clone,
+    Eq,
+    PartialEq,
+    Serialize,
+    Deserialize,
+    Decode,
+    Encode,
+    FromVariant,
+    Error,
+    IntoSchema,
+)]
+pub enum RejectionReason {
+    /// The reason for rejecting the block.
+    #[error("Block was rejected")]
+    Block(#[source] BlockRejectionReason),
+    /// The reason for rejecting transaction.
+    #[error("Transaction was rejected")]
+    Transaction(#[source] TransactionRejectionReason),
+}
+
 /// The prelude re-exports most commonly used traits, structs and macros from this module.
 pub mod prelude {
     pub use super::{
-        Payload, PendingTransactions, RejectedTransaction, Transaction, TransactionValue, Txn,
+        BlockRejectionReason, InstructionExecutionFail, NotPermittedFail, Payload,
+        PendingTransactions, RejectedTransaction, RejectionReason, Transaction,
+        TransactionRejectionReason, TransactionValue, Txn, UnsatisfiedSignatureConditionFail,
         ValidTransaction, VersionedPendingTransactions, VersionedRejectedTransaction,
         VersionedTransaction, VersionedValidTransaction,
     };
